% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{agrupar_duplicados}
\alias{agrupar_duplicados}
\title{funcion para agrupar los duplicados. ¿Quien está duplicado con quien?}
\usage{
agrupar_duplicados(data, n = 31, m = 1)
}
\arguments{
\item{data}{matriz de datos}

\item{n}{posición final de periodo (depende de número de días del mes y de si tenemos varios datos diarios o solo uno)}

\item{m}{posición de inicio del periodo}
}
\value{
lista de matrices de duplicados
}
\description{
funcion para agrupar los duplicados. ¿Quien está duplicado con quien?
}
