% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{agrupar_duplicados_offset}
\alias{agrupar_duplicados_offset}
\title{agrupa la matriz de duplicados}
\usage{
agrupar_duplicados_offset(x, a)
}
\arguments{
\item{x}{matriz de datos}

\item{a}{matriz de duplicados}
}
\value{
lista de matrices
}
\description{
agrupa la matriz de duplicados
}
