% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{agrupar_metadatos_duplicados}
\alias{agrupar_metadatos_duplicados}
\title{reorganiza metadatos generados durante la ejecución}
\usage{
agrupar_metadatos_duplicados(metadatos, var, pasada = 1)
}
\arguments{
\item{metadatos}{matriz de metadatos}

\item{var}{tipo de dato (tmin, pr...)}

\item{pasada}{distinguimos entre la primera y segunda pasada para guardar algunos datos}
}
\value{
None
}
\description{
reorganiza metadatos generados durante la ejecución
}
