% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{buscar_duplicados}
\alias{buscar_duplicados}
\title{busca los meses duplicados entre dia 1 y dia 'n', meses duplicados pero que no tienen el mismo número de dias.}
\usage{
buscar_duplicados(data, meta, nz = 10, n, var)
}
\arguments{
\item{data}{matriz de datos}

\item{meta}{matriz de metadatos de los datos}

\item{nz}{dias con dato necesarios para hacer estas comprobaciones.}

\item{n}{numero de dias que queremos comprobar (de 28 a 31, para comprobar todos los meses posibles)}

\item{var}{tipo de dato (tmin, pr...)}
}
\value{
lista de matriz de datos y de metadatos
}
\description{
busca los meses duplicados entre dia 1 y dia 'n', meses duplicados pero que no tienen el mismo número de dias.
}
