% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{cien_entre_nas}
\alias{cien_entre_nas}
\title{Si hay mas de 'b' dias de 100% despues (o justo antes) de una cadena de 'a' nas consideramos error el 100 como valor erroneo}
\usage{
cien_entre_nas(data, metadatos, A = 8, B = 6)
}
\arguments{
\item{data}{Matriz de datos}

\item{metadatos}{metadatos de los datos}

\item{A}{Longitud mínima de la cadena de NAs}

\item{B}{Longitud mínima de la cadena de 100's}
}
\value{
Metadatos de los datos
}
\description{
Si hay mas de 'b' dias de 100% despues (o justo antes) de una cadena de 'a' nas consideramos error el 100 como valor erroneo
}
