% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{combinar_metadatos_consecutivos}
\alias{combinar_metadatos_consecutivos}
\title{organizar metadatos de consecutivos}
\usage{
combinar_metadatos_consecutivos(
  metadatos_a,
  metadatos_b,
  x = NULL,
  var = C_MAX
)
}
\arguments{
\item{metadatos_a}{matriz de metadatos}

\item{metadatos_b}{matriz de metadatos}

\item{x}{resolucion de los datos}

\item{var}{tipo de datos (tmin, pr...)}
}
\value{
matriz de metadatos
}
\description{
organizar metadatos de consecutivos
}
