% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{compatibles_row_col}
\alias{compatibles_row_col}
\title{Elimina filas y columnas no compartidas entre las 2 matrices que recibe}
\usage{
compatibles_row_col(data1, data2)
}
\arguments{
\item{data1}{Matriz que se devolverá una vez eliminadas filas y columnas no presentes en el segundo parámetro}

\item{data2}{Matriz auxiliar con la que comparar filas y columnas}
}
\value{
Matriz de datos
}
\description{
Elimina filas y columnas no compartidas entre las 2 matrices que recibe
}
