% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{comprobar_consecutivos_insolacion}
\alias{comprobar_consecutivos_insolacion}
\title{comprobar que hay valores por debajo del 70% del m치ximo para la fecha y posici칩n del dato}
\usage{
comprobar_consecutivos_insolacion(datos, metadatos)
}
\arguments{
\item{datos}{matriz de datos}

\item{metadatos}{matriz de metadatos de los datos}
}
\value{
matriz de metadatos
}
\description{
comprobar que hay valores por debajo del 70% del m치ximo para la fecha y posici칩n del dato
}
