% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{consecutivos}
\alias{consecutivos}
\title{deteccion de datos consecutivos iguales}
\usage{
consecutivos(datos, metadatos, C = 7)
}
\arguments{
\item{datos}{matriz de datos}

\item{metadatos}{matriz de metadatos de los datos}

\item{C}{cuantos días consecutivos buscamos, 7 o 14}
}
\description{
deteccion de datos consecutivos iguales
}
