% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controles_hr.R
\name{contol_hr}
\alias{contol_hr}
\title{Tratamiento de datos de rh, final de la revisión de datos diarios}
\usage{
contol_hr(RH, metadatos_originales_RH)
}
\arguments{
\item{RH}{datos}

\item{metadatos_originales_RH}{metadatos_originales}
}
\value{
None
}
\description{
Tratamiento de datos de rh, final de la revisión de datos diarios
}
