% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controles_in.R
\name{contol_ins}
\alias{contol_ins}
\title{Tratamiento de datos de ins, final de la revisión de datos diarios}
\usage{
contol_ins(ins, metadatos_originales_ins)
}
\arguments{
\item{ins}{datos}

\item{metadatos_originales_ins}{metadatos_originales}
}
\description{
Tratamiento de datos de ins, final de la revisión de datos diarios
}
