% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controles_pr.R
\name{contol_pp}
\alias{contol_pp}
\title{Tratamiento de datos de pp, final de la revisión de datos diarios}
\usage{
contol_pp(Pre, metadatos_originales_pre, dist)
}
\arguments{
\item{dist}{}
}
\description{
Tratamiento de datos de pp, final de la revisión de datos diarios
}
