% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controles_r.R
\name{contol_ra}
\alias{contol_ra}
\title{Tratamiento de datos de ra, final de la revisión de datos diarios}
\usage{
contol_ra(Ra, metadatos_originales_ra)
}
\arguments{
\item{Ra}{datos}

\item{metadatos_originales_ra}{metadatos_originales}
}
\description{
Tratamiento de datos de ra, final de la revisión de datos diarios
}
