% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_1.R
\name{contol_temp1}
\alias{contol_temp1}
\title{Tratamiento de los datos diarios leídos, 1º parte}
\usage{
contol_temp1(min, dist, namevar)
}
\arguments{
\item{min}{datos diarios de la variable}

\item{namevar}{variable a tratar}
}
\value{
datos de la variable tratados y metadatos originales
}
\description{
Tratamiento de los datos diarios leídos, 1º parte
}
