% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controles_t.R
\name{contol_temp15}
\alias{contol_temp15}
\title{Tratamiento de datos de t, una parte}
\usage{
contol_temp15(min, namevar)
}
\arguments{
\item{min}{datos}

\item{namevar}{nombre variable}
}
\description{
Tratamiento de datos de t, una parte
}
