% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controles_t.R
\name{contol_temp2}
\alias{contol_temp2}
\title{Tratamiento de datos de t, una parte}
\usage{
contol_temp2(min, metadatos_originales_min_ancho, namevar)
}
\arguments{
\item{min}{datos}

\item{metadatos_originales_min_ancho}{metadatos_originales}

\item{namevar}{nombre variable}
}
\description{
Tratamiento de datos de t, una parte
}
