% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controles_t.R
\name{contol_temp3}
\alias{contol_temp3}
\title{Tratamiento de datos de t, final de la revisión de datos diarios}
\usage{
contol_temp3(min, namevar)
}
\arguments{
\item{min}{datos}

\item{namevar}{nombre variable}
}
\description{
Tratamiento de datos de t, final de la revisión de datos diarios
}
