% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_1.R
\name{controles}
\alias{controles}
\title{Control de datos de mtomas sobre las variable indicada}
\usage{
controles(var)
}
\arguments{
\item{var}{variable a tratar}
}
\value{
None
}
\description{
Control de datos de mtomas sobre las variable indicada
}
