% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{correctUnits}
\alias{correctUnits}
\title{Correct units of a weather series}
\usage{
correctUnits(series, type)
}
\arguments{
\item{series}{a numeric vector containing the values of the weather series to be corrected}

\item{type}{a character string indicating the type of weather series, which can be one of 
"w" (wind speed), "hr" (relative humidity), "pr" (precipitation), "in" (insolation),
"r" (radiation), "t" (temperature).}
}
\value{
a numeric vector with the corrected values of the input series
}
\description{
This function corrects the units of a weather series based on its type.
}
