% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{crear_metadatos_originales}
\alias{crear_metadatos_originales}
\title{creamos metadatos originales para temperatura}
\usage{
crear_metadatos_originales(datos)
}
\arguments{
\item{datos}{matriz de datos}
}
\value{
Matriz de datos
}
\description{
creamos metadatos originales para temperatura
}
