% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{daily_range}
\alias{daily_range}
\title{Guarda en los metadatos los días en los que el rango de temperatura supera un límite}
\usage{
daily_range(data_min, data_max, metadatos_min, metadatos_max, th_dr = 350)
}
\arguments{
\item{data_min}{Matriz de datos de tmin}

\item{data_max}{Matriz de datos de tmax}

\item{metadatos_min}{Matriz de metadatos de tmin}

\item{metadatos_max}{Matriz de metadatos de tmax}

\item{th_dr}{Diferencia permitida entre tmin y tmax}
}
\value{
None
}
\description{
Guarda en los metadatos los días en los que el rango de temperatura supera un límite
}
