% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{decadal_duplicated}
\alias{decadal_duplicated}
\title{detectar decenas climaticas duplicadas}
\usage{
decadal_duplicated(data, meta, var, dec = 1, data_source = C_AEMET)
}
\arguments{
\item{data}{matriz de datos}

\item{meta}{matriz de metadatos de los datos}

\item{var}{tipo de datos (tmin, pr...)}

\item{dec}{si busca en la 1, 2 o 3 decena del mes}

\item{data_source}{fuente de los datos (AEMET, SIAR)}
}
\value{
lista de matrices de duplicados y metadatos
}
\description{
detectar decenas climaticas duplicadas
}
