% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{delete_na}
\alias{delete_na}
\title{Elimina filas sin ningún valor distinto de NA y de -999}
\usage{
delete_na(data)
}
\arguments{
\item{data}{Matriz que se devolverá una vez eliminadas que no aportan información}
}
\value{
Matriz de datos
}
\description{
Elimina filas sin ningún valor distinto de NA y de -999
}
