% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{delete_no_dist}
\alias{delete_no_dist}
\title{Elimina filas sin datos sobre su posici칩n en la matriz de distancias}
\usage{
delete_no_dist(data, dist)
}
\arguments{
\item{data}{Matriz que se devolver치 una vez eliminadas filas de las que no conocemos distancias a la estaci칩n}

\item{dist}{Matriz de distancias}
}
\value{
Matriz de datos
}
\description{
Elimina filas sin datos sobre su posici칩n en la matriz de distancias
}
