% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{deteccion_aberrantes}
\alias{deteccion_aberrantes}
\title{Marca en los metadatos los días con fechas aberrantes}
\usage{
deteccion_aberrantes(data, metadatos, var, vart = NULL)
}
\arguments{
\item{data}{matriz de datos}

\item{metadatos}{matriz de metadatos}

\item{var}{tipo de datos (tmin, pr...)}

\item{vart}{Hora del día de los datos (no para todos los var)}
}
\value{
Matriz de metadatos
}
\description{
Marca en los metadatos los días con fechas aberrantes
}
