% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{deteccion_decenas_duplicadas}
\alias{deteccion_decenas_duplicadas}
\title{detecta decenas duplicadas dentro de los meses de datos}
\usage{
deteccion_decenas_duplicadas(data, metadatos, var, data_source = C_AEMET)
}
\arguments{
\item{data}{matriz de datos}

\item{metadatos}{matriz de metadatos de los datos}

\item{var}{tipo de datos (tmin, pr...)}

\item{data_source}{fuente de los datos (AEMET, SIAR)}
}
\value{
lista de matriz de metadatos y una lista de duplicados por decenas de días del mes
}
\description{
detecta decenas duplicadas dentro de los meses de datos
}
