% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{deteccion_duplicados}
\alias{deteccion_duplicados}
\title{funcion para detectar duplicados}
\usage{
deteccion_duplicados(datos, metadatos, var, data_source = C_AEMET)
}
\arguments{
\item{datos}{matriz de datos}

\item{metadatos}{matriz de metadatos de los datos}

\item{var}{tipo de datos (tmin, pr...)}

\item{data_source}{fuente de los datos (AEMET o SIAR)}
}
\value{
duplicados es una lista que contiene 4 elementos. Los duplicados a 28, 29, 30 y 31 dias; metadatos_meses_duplicados_detectados vale '1' en duplicados, '0' en originales, 'NA' en no disponibles
}
\description{
funcion para detectar duplicados
}
