% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{deteccion_duplicados_25_dias}
\alias{deteccion_duplicados_25_dias}
\title{deteccion de meses con 25 dias identicos}
\usage{
deteccion_duplicados_25_dias(data, metadatos, var, dist, data_source = C_AEMET)
}
\arguments{
\item{metadatos}{matriz de metadatos correspondientes a los datos}

\item{var}{tipo de datos (tmin, pr...)}

\item{dist}{matriz de distancias}

\item{data_source}{Distinguir si tratamos datos de la AEMET o del SIAR}

\item{datos}{matriz de datos}
}
\value{
None
}
\description{
deteccion de meses con 25 dias identicos
}
