% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{detectar_consecutivos}
\alias{detectar_consecutivos}
\title{deteccion de datos consecutivos}
\usage{
detectar_consecutivos(datos, metadatos, var, vart = NULL)
}
\arguments{
\item{metadatos}{matriz de metadatos}

\item{var}{tipo de datos (tmin, pr...)}

\item{vart}{Hora del día de los datos (no para todos los var)}

\item{data}{matriz de datos}
}
\value{
matriz de metadatos o none
}
\description{
deteccion de datos consecutivos
}
