% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{distancias}
\alias{distancias}
\title{Crea archivo de distancias}
\usage{
distancias(a, var, data_source = C_AEMET)
}
\arguments{
\item{a}{directorio en el que están los ficheros}

\item{var}{tipo de datos (tmin, pr...)}

\item{data_source}{fuente de los datos (AEMET o SIAR)}
}
\value{
data frame de datos
}
\description{
Crea archivo de distancias
}
