% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{eliminar_datos}
\alias{eliminar_datos}
\title{Pone NAs en la matriz de datos si así lo indican los valores de los metadatos}
\usage{
eliminar_datos(data, metadata)
}
\arguments{
\item{data}{matriz de datos}

\item{metadata}{matriz de metadatos de los datos}
}
\value{
matriz de datos
}
\description{
Pone NAs en la matriz de datos si así lo indican los valores de los metadatos
}
