% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{ficheroDistanciasLeer}
\alias{ficheroDistanciasLeer}
\title{Lee archivos de distancias y homogeneiza las coordenadas}
\usage{
ficheroDistanciasLeer()
}
\value{
matrizde datos
}
\description{
Lee archivos de distancias y homogeneiza las coordenadas
}
