% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{ficheroDistanciasLeer1}
\alias{ficheroDistanciasLeer1}
\title{Lee archivos de distancias y homogeneiza las coordenadas}
\usage{
ficheroDistanciasLeer1()
}
\value{
matriz de datos
}
\description{
Lee archivos de distancias y homogeneiza las coordenadas
}
