% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{filter_dec_dupl}
\alias{filter_dec_dupl}
\title{filtrar decenas duplicadas. Si la deteccion es entre estaciones diferentes, tiene que ser mismo mes y mismo año}
\usage{
filter_dec_dupl(dat)
}
\arguments{
\item{dat}{lista de matrices de duplicados}
}
\value{
lista de matrices de duplicados que cumplen la condición
}
\description{
filtrar decenas duplicadas. Si la deteccion es entre estaciones diferentes, tiene que ser mismo mes y mismo año
}
