% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{from_subdaily_to_daily}
\alias{from_subdaily_to_daily}
\title{pasa de datos subdiarios a datos diarios}
\usage{
from_subdaily_to_daily(dat07, dat13, dat18)
}
\arguments{
\item{dat07}{matriz de datos a las 7}

\item{dat13}{matriz de datos a las 13}

\item{dat18}{matriz de datos a las 18}
}
\value{
matriz de datos diarios
}
\description{
pasa de datos subdiarios a datos diarios
}
