% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{intradecadal_duplicated}
\alias{intradecadal_duplicated}
\title{para detectar duplicidades entre decenas climaticas de un mismo mes; metadatos de salida vale '1' para los casos que estan duplicados, '0' para datos originales y 'NA'}
\usage{
intradecadal_duplicated(data, meta, var, data_source = C_AEMET)
}
\arguments{
\item{data}{matriz de datos}

\item{meta}{matriz de metadatos de los datos}

\item{var}{tipo de datos (tmin, pr...)}

\item{data_source}{data_source fuente de los datos (AEMET, SIAR)}
}
\value{
lista de matrices de datos duplicados y metadatos
}
\description{
para detectar duplicidades entre decenas climaticas de un mismo mes; metadatos de salida vale '1' para los casos que estan duplicados, '0' para datos originales y 'NA'
}
