% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch.all.controls}
\alias{launch.all.controls}
\title{Launch all quality controls}
\usage{
launch.all.controls()
}
\value{
No return value.
}
\description{
This function launches all quality controls for the variables: wind speed, relative humidity,
precipitation, sunshine duration, radiation and temperature.
}
