% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch.controls}
\alias{launch.controls}
\title{Apply quality controls to specified variables.}
\usage{
launch.controls(vars)
}
\arguments{
\item{vars}{A character vector containing the variable names to apply quality controls to.}
}
\description{
This function applies quality controls to the specified variables.
}
