% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{lectura_datos_aemet}
\alias{lectura_datos_aemet}
\title{Lee datos del AEMET desde los ficheros de origen y los devuelve}
\usage{
lectura_datos_aemet(a, var)
}
\arguments{
\item{a}{directorio en el que están los ficheros}

\item{var}{tipo de datos (tmin, pr...)}

\item{data_source}{fuente de los datos (AEMET o SIAR)}
}
\value{
Matriz o lista de matrices de datos
}
\description{
Lee datos del AEMET desde los ficheros de origen y los devuelve
}
