% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{lectura_datos_siar}
\alias{lectura_datos_siar}
\title{Lee datos del SIAR desde los ficheros de origen y los devuelve}
\usage{
lectura_datos_siar(a, var)
}
\arguments{
\item{a}{directorio en el que están los ficheros}

\item{var}{tipo de datos (tmin, pr...)}
}
\value{
Matriz o lista de matrices de datos
}
\description{
Lee datos del SIAR desde los ficheros de origen y los devuelve
}
