% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{load.data}
\alias{load.data}
\title{Load data from a file in a specified directory}
\usage{
load.data(file)
}
\arguments{
\item{file}{Name of the file to load data from}

\item{dataOutFiles}{Directory where the file is located}
}
\value{
Objects loaded from the file
}
\description{
Load data from a file in a specified directory
}
