% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{mala_codificacion_temperatura}
\alias{mala_codificacion_temperatura}
\title{Detecta mala codificación en las temperaturas}
\usage{
mala_codificacion_temperatura(datos, metadatos, var)
}
\arguments{
\item{datos}{matriz de datos}

\item{metadatos}{matriz de metadatos}

\item{var}{tipo de datos (tmin, pr...)}
}
\value{
matriz de metadatos o none
}
\description{
Detecta mala codificación en las temperaturas
}
