% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{max_mobil}
\alias{max_mobil}
\title{Control específico para el numero de horas de sol. Se busca el maximo de horas
de sol dentro de una ventana móbil. Útil para detectar anomalías relacionadas
con la codificación. Hay casos en las series de AEMET en los que existen
codificaciones erróneas (aparentemente divididas por 10 respecto a su valor
observado.}
\usage{
max_mobil(a, b, C = 15, D = 20)
}
\arguments{
\item{a}{matriz de datos}

\item{b}{metadatos de los datos}

\item{C}{Tamaño de la ventana móvil (parece mal usado, ya que se usa directamente su valor por defecto en algunos sitios)}

\item{D}{Parámetro para los cálculos, máximo de horas de insolación}
}
\value{
Metadatos de los datos
}
\description{
Control específico para el numero de horas de sol. Se busca el maximo de horas
de sol dentro de una ventana móbil. Útil para detectar anomalías relacionadas
con la codificación. Hay casos en las series de AEMET en los que existen
codificaciones erróneas (aparentemente divididas por 10 respecto a su valor
observado.
}
