% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{maximo_teorico}
\alias{maximo_teorico}
\title{Máximo teórico de horas de sol}
\usage{
maximo_teorico(month, lat, var = C_INS, dias = NULL)
}
\arguments{
\item{month}{mes}

\item{lat}{latitud}

\item{var}{tipo de datos (insolació o radiación)}

\item{dias}{día del mes}
}
\value{
valor
}
\description{
Máximo teórico de horas de sol
}
