% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{meses_duplicados}
\alias{meses_duplicados}
\title{deteccion de meses duplicados}
\usage{
meses_duplicados(
  datos,
  metadatos,
  dist,
  var,
  pasada = 1,
  data_source = C_AEMET
)
}
\arguments{
\item{datos}{matriz de datos}

\item{metadatos}{matriz de metadatos correspondientes a los datos}

\item{dist}{matriz de distancias}

\item{var}{tipo de datos (tmin, pr...)}

\item{pasada}{La pr distingue entre la pasada 1 y la 2 a la hora de guardar algunos datos}

\item{data_source}{Distinguir si tratamos datos de la AEMET o del SIAR}
}
\value{
None
}
\description{
deteccion de meses duplicados
}
