% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controles_in.R
\name{metadata_ins}
\alias{metadata_ins}
\title{Preparación de los metadatos de ins para su guardado}
\usage{
metadata_ins()
}
\value{
None
}
\description{
Preparación de los metadatos de ins para su guardado
}
