% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controles_t.R
\name{metadata_temp}
\alias{metadata_temp}
\title{Preparación de los metadatos de temp para su guardado}
\usage{
metadata_temp(namevar)
}
\value{
None
}
\description{
Preparación de los metadatos de temp para su guardado
}
