% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{min_sup_eq_max}
\alias{min_sup_eq_max}
\title{Anota en los metadatos si la tmin es mayor o igual que la tmax}
\usage{
min_sup_eq_max(data_min, data_max, metadatos_min, metadatos_max)
}
\arguments{
\item{data_min}{Matriz de datos de tmin}

\item{data_max}{Matriz de datos de tmax}

\item{metadatos_min}{Matriz de metadatos de tmin}

\item{metadatos_max}{Matriz de metadatos de tmax}
}
\value{
None
}
\description{
Anota en los metadatos si la tmin es mayor o igual que la tmax
}
