% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{n_dias_duplicados}
\alias{n_dias_duplicados}
\title{Si hay más de 6 días con lluvia idéntica entre meses consecutivos se considera que el mes está duplicado}
\usage{
n_dias_duplicados(data, meta)
}
\arguments{
\item{data}{data frame de datos}

\item{metadata}{frame de metadatos correspondientes a los datos}
}
\value{
None
}
\description{
Si hay más de 6 días con lluvia idéntica entre meses consecutivos se considera que el mes está duplicado
}
