% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{obtener_resolucion}
\alias{obtener_resolucion}
\title{obtenemos la resolucion de los datos}
\usage{
obtener_resolucion(datos, metadatos)
}
\arguments{
\item{datos}{matriz de datos}

\item{metadatos}{matriz de metadatos correspondientes a los datos}
}
\value{
matriz de metadatos
}
\description{
obtenemos la resolucion de los datos
}
