% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{offset_duplicated_rain}
\alias{offset_duplicated_rain}
\title{busca meses con las mismas cantidades de lluvia, aunque no esten en el mismo orden}
\usage{
offset_duplicated_rain(data, meta, pasada)
}
\arguments{
\item{data}{matriz de datos}

\item{meta}{matriz de metadatos correspodientes a los datos de la matriz}

\item{pasada}{distinguimos entre la primera y segunda pasada para guardar algunos datos}
}
\value{
None
}
\description{
busca meses con las mismas cantidades de lluvia, aunque no esten en el mismo orden
}
