% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{poca_variacion_bucle}
\alias{poca_variacion_bucle}
\title{Busca datos de hr con poca variabilidad}
\usage{
poca_variacion_bucle(x_07, x_13, x_18, C = 10, E = 5, G = T)
}
\arguments{
\item{x_07}{matriz de datos a las 7 horas}

\item{x_13}{matriz de datos a las 13 horas}

\item{x_18}{matriz de datos a las 18 horas}

\item{C}{Tamaño de la ventana móvil}

\item{E}{variabilidad}

\item{G}{si los datos máximos deben estar por debajo o por encima de 95}
}
\value{
Matriz de metadatos
}
\description{
Busca datos de hr con poca variabilidad
}
