% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{prepare.data}
\alias{prepare.data}
\title{Prepare data for analysis}
\usage{
prepare.data(name.type)
}
\arguments{
\item{name.type}{character string indicating the type of data to prepare}

\item{dataOutFiles}{character string indicating the path to the directory where the output data will be stored}

\item{year.new}{integer indicating the year of the data}

\item{data_source}{character string indicating the source of the data}
}
\value{
None
}
\description{
This function prepares the data for analysis by reading in the processed data
from a file, correcting the units, and writing the data to new files with the
appropriate format for analysis.
}
