% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{qc.apply}
\alias{qc.apply}
\title{qc.apply}
\usage{
qc.apply(vars, output.folder = "new_all", data.source = "AEMET")
}
\arguments{
\item{vars}{Character vector of variable types to apply QC to}

\item{output.folder}{Character string of the output data folder path}

\item{data.source}{Character string indicating the data source (default is "AEMET")}

\item{input.folder}{Character string of the input data folder path}
}
\value{
None
}
\description{
Apply quality control to meteorological data
}
