% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{reformat}
\alias{reformat}
\title{reformateamos datos a formato ancho}
\usage{
reformat(datos)
}
\arguments{
\item{datos}{matriz de datos}
}
\value{
data frame de de los datos con el nuevo formato
}
\description{
reformateamos datos a formato ancho
}
