% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_1.R
\name{rename_metadata}
\alias{rename_metadata}
\title{Renombrado de metadatos}
\usage{
rename_metadata(namevar)
}
\arguments{
\item{namevar}{variable a tratar}
}
\value{
None
}
\description{
Renombrado de metadatos
}
