% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{save.data}
\alias{save.data}
\title{Save data to a specified file}
\usage{
save.data(..., file)
}
\arguments{
\item{...}{objects to be saved}

\item{file}{a character string specifying the file name}

\item{dataOutFiles}{a character string specifying the directory where the file should be saved}
}
\value{
This function saves the data to the specified file.
}
\description{
This function saves data to a specified file in a given directory.
}
