% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{special_character}
\alias{special_character}
\title{Function to delete -4 from the database previous to any other control
TS (TIME-SCALE).
TS=='D' (DAILY) ALL -4 and the acumulated rainfall is deleted
TS=='M' (MONTHLY) ALL -4 are deleted but the acumulated monthly data is preserved}
\usage{
special_character(data, bucle = T, TS = "D")
}
\arguments{
\item{data}{Array de datos}

\item{bucle}{True para usar bucle en otro caso usar apply, elbucle ahorra RAM y el apply tiempo}

\item{TS}{D si queremos eliminar todos los -4}
}
\value{
Lista con los datos y los matedatos
}
\description{
Function to delete -4 from the database previous to any other control
TS (TIME-SCALE).
TS=='D' (DAILY) ALL -4 and the acumulated rainfall is deleted
TS=='M' (MONTHLY) ALL -4 are deleted but the acumulated monthly data is preserved
}
