% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{tratamiento_decadas}
\alias{tratamiento_decadas}
\title{tratar decadas duplicadas. Si tenemos una decena duplicada solo eliminamos esa decena, si tenemos
dos decenas duplicadas, borramos ambas}
\usage{
tratamiento_decadas(s, b, dec, m = 1, n = 10)
}
\arguments{
\item{s}{lista de datos duplicados}

\item{b}{matriz de metadatos de los datos}

\item{dec}{encontradas 1 decena duplicada o 2 decenas duplicadas en el mes}

\item{m}{posición de inicio}

\item{n}{posición de fin (según númoer de datos diarios y numero de días del mes)}
}
\value{
matriz de metadatos de los datos
}
\description{
tratar decadas duplicadas. Si tenemos una decena duplicada solo eliminamos esa decena, si tenemos
dos decenas duplicadas, borramos ambas
}
