% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{tratamiento_duplicados}
\alias{tratamiento_duplicados}
\title{crea una lista de 4 posiciones (28, 29, 30 y 31 días) que tiene valores: '2' (eliminar), '3' (conservar), '4' (conservar por cercanía espacial) eliminar contiene las filas de los elementos que se tienen que eliminar}
\usage{
tratamiento_duplicados(
  datos,
  duplicados,
  metadatos,
  dist,
  var,
  data_source = C_AEMET
)
}
\arguments{
\item{datos}{matriz de datos}

\item{metadatos}{matriz de metadatos de los datos}

\item{dist}{matriz de distancias}

\item{var}{tipo de dato (tmin, pr...)}

\item{data_source}{matriz de metadatos de los datos}
}
\value{
lista de matrices de datos a eliminar y metadatos
}
\description{
crea una lista de 4 posiciones (28, 29, 30 y 31 días) que tiene valores: '2' (eliminar), '3' (conservar), '4' (conservar por cercanía espacial) eliminar contiene las filas de los elementos que se tienen que eliminar
}
