% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{tratamiento_unicos}
\alias{tratamiento_unicos}
\title{funcion para decdidir que hacemos con duplicados. ¿Borramos o conservamos por distancia?}
\usage{
tratamiento_unicos(x, s, b, d, n, cons = F)
}
\arguments{
\item{x}{matriz de datos}

\item{b}{matriz de metadatos de los datos}

\item{d}{matriz de distancias}

\item{n}{1, 2, 3 o 4 que equivalen a '28', '29', '30' o '31' días}

\item{cons}{TRUE para eliminar duplicados consecutivos}
}
\value{
matriz de metadatos de los datos
}
\description{
funcion para decdidir que hacemos con duplicados. ¿Borramos o conservamos por distancia?
}
