% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{tratamiento_unicos_25}
\alias{tratamiento_unicos_25}
\title{tratamiento de meses con 25 días identicos}
\usage{
tratamiento_unicos_25(x, s, b, d)
}
\arguments{
\item{x}{matriz de datos}

\item{s}{lista de matrices de duplicados}

\item{b}{matriz de metadatos}

\item{d}{matriz de distancias}
}
\value{
matriz de metadatos
}
\description{
tratamiento de meses con 25 días identicos
}
