% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC_2.R
\name{tratar_decenas_duplicadas}
\alias{tratar_decenas_duplicadas}
\title{devuelve '2' el mes tiene dos decenas duplicadas, hay que eliminar el mes entero. '5' el mes tiene una única decena duplicada. Hay que eliminar únicamente esa decena.}
\usage{
tratar_decenas_duplicadas(dec_dupl, metadatos, var, data_source = C_AEMET)
}
\arguments{
\item{dec_dupl}{lista de datos duplicados}

\item{metadatos}{matriz de metadatos de los datos}

\item{var}{tipo de datos (tmin, pr...)}

\item{data_source}{fuente de los datos (AEMET, SIAR)}
}
\value{
matriz de metadatos
}
\description{
devuelve '2' el mes tiene dos decenas duplicadas, hay que eliminar el mes entero. '5' el mes tiene una única decena duplicada. Hay que eliminar únicamente esa decena.
}
