% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{year.new.dir.calc}
\alias{year.new.dir.calc}
\title{Calculate the directory name for a given year or day value}
\usage{
year.new.dir.calc(year.new)
}
\arguments{
\item{year.new}{A numeric or character value representing a year or day.}
}
\value{
A character string representing the directory name.
}
\description{
This function calculates the directory name for a given year or day value. If the value
is a year, the directory name will be the same as the year value. If the value is a day,
the directory name will be "data_QC". If the value is missing or NA, the directory name
will also be "data_QC".
}
\keyword{day}
\keyword{directory}
\keyword{name,}
\keyword{year,}
