% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{zeros_aislados}
\alias{zeros_aislados}
\title{Busca falsos 0's}
\usage{
zeros_aislados(datos, metadatos, var)
}
\arguments{
\item{datos}{Matriz de datos}

\item{metadatos}{Matriz de metadatos}

\item{var}{tipo de datos (pr, tmin...)}
}
\value{
None
}
\description{
Busca falsos 0's
}
