% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R
\name{zeros_dobles}
\alias{zeros_dobles}
\title{0's que estan tanto en las minimas como en las maximas}
\usage{
zeros_dobles(dat1, dat2)
}
\arguments{
\item{dat1}{Matriz de tmin o tmax}

\item{dat2}{Matriz de tmin o tmax}
}
\value{
None
}
\description{
0's que estan tanto en las minimas como en las maximas
}
\examples{
dat1=min; dat2=max
}
