% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Controles_w.R
\name{contol_w}
\alias{contol_w}
\title{Tratamiento de datos de W, final de la revisión de datos diarios.
En primer lugar se agrega de datos sub-diarios (lecturas a las 00, 07,
13 y 18 horas) a datos diarios. Luego se detectan datos repetidos,
datos aberrantes, y los propios controles de calidad de AEMET.}
\usage{
contol_w(U10, metadatos_originales_U10)
}
\arguments{
\item{U10}{datos}

\item{metadatos_originales_U10}{metadatos_originales}
}
\description{
Tratamiento de datos de W, final de la revisión de datos diarios.
En primer lugar se agrega de datos sub-diarios (lecturas a las 00, 07,
13 y 18 horas) a datos diarios. Luego se detectan datos repetidos,
datos aberrantes, y los propios controles de calidad de AEMET.
}
