% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones_QC.R, R/pruebas1.R
\name{ficheroDistanciasLeer}
\alias{ficheroDistanciasLeer}
\title{Lee archivos de distancias y homogeneiza las coordenadas}
\usage{
ficheroDistanciasLeer()

ficheroDistanciasLeer()
}
\value{
matriz de datos

matriz de datos
}
\description{
Lee archivos de distancias y homogeneiza las coordenadas

Lee archivos de distancias y homogeneiza las coordenadas
}
