% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Termo.R
\docType{data}
\name{Termo}
\alias{Termo}
\title{Temperature records in Spain}
\format{
A dataframe with 6083 rows and 72 variables.
}
\usage{
data(Termo)
}
\description{
This is a little sample of data which contains information about the temperature measured by six weather stations in the Balearic Islands (Spain). The dataframe is formed by 6083 observations and 72 variables corresponding to more than 5 years (62 months). The temperature is measured in degrees Celsius multiplied by ten to avoid decimals.
}
\section{Details}{

 The dataframe contains the following variables:
 \describe{
   \itemize{
     \item{INDICATIVO: }{The code of every weather station}
     \item{AÑO: }{The year of the recording}
     \item{MES: }{The month of the recording}
     \item{NOMBRE: }{The name of the weather station}
     \item{ALTITUD: }{The elevation of each station}
     \item{C_X: }{The X UTM coordinates of the weather station}
     \item{C_Y: }{The Y UTM coordinates of the weather station}
     \item{NOM_PROV: }{The name of the province where the station is located}
     \item{LONGITUD: }{The longitude coordinates of the weather station}
     \item{LATITUD: }{The longitude coordinates of the weather station}
     \item{TMAXn: }{The maximum temperature recorded the "n" day of the month}
     \item{TMINn: }{The minimum temperature recorded the "n" day of the month}
   }
 }
 
 \source{
   The data has been extracted from a small dataset made as example coming from the Spanish Meteorological Agency. For more information, visit the following link:
     \url{https://www.aemet.es/es/portada}
 }
}

\examples{
my_dataframe <- qualityControl::Termo

}
\keyword{datasets}
