% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Termo.R
\docType{data}
\name{Termo}
\alias{Termo}
\title{Temperature records in Spain}
\format{
A dataframe with 6083 rows and 72 variables.
}
\usage{
data(Termo)
}
\description{
This is a little sample of data which contains information about the temperature measured by six weather stations in Spain. The dataframe is formed by 6083 observations and 72 variables corresponding to the year 2022. The temperature is measured in degrees Celsius multiplied by ten to avoid decimals. It is worth mentioning that these data are invented, they do not correspond to real data. The aim is simply to have an example dataframe with similar characteristics to a real one.
}
\section{Details}{

 The dataframe contains the following variables:
 \describe{
   \itemize{
     \item{INDICATIVO: }{The code of every weather station}
     \item{AÑO: }{The year of the recording}
     \item{MES: }{The month of the recording}
     \item{NOMBRE: }{The name of the weather station}
     \item{ALTITUD: }{The elevation of each station}
     \item{C_X: }{The X UTM coordinates of the weather station. The reference system is ETRS89. Data are measured in metres.}
     \item{C_Y: }{The Y UTM coordinates of the weather station. The reference system is ETRS89. Data are measured in metres.}
     \item{NOM_PROV: }{The name of the province where the station is located}
     \item{LONGITUD: }{The longitude coordinates of the weather station}
     \item{LATITUD: }{The longitude coordinates of the weather station}
     \item{TMAXn: }{The maximum temperature recorded the "n" day of the month}
     \item{TMINn: }{The minimum temperature recorded the "n" day of the month}
   }
 }
 
 \source{
   TThese data are a simulation of real data from AEMET, carried out by the LCSC (CSIC) in the context of a joint project on research and analysis of climate data. For more information, visit the following link:
     \url{https://lcsc.csic.es/es/lcsc/}
 }
}

\examples{
my_dataframe <- qualityControl::Termo

}
\keyword{datasets}
